Version 9 final project
I added a second visualization to the eda and descriptions for both visualizations. I added a written out methods section per the assignment instructions. I added a summary results figure at the end of the analysis before the results interpretation. I added a little bit to the results interpretation. Did not touch the discussion.